requirements.txt
langchain==0.2.14
langchain-community==0.2.12
langchain-core==0.2.33
langchain-openai==0.1.23
pypdf==4.2.0
sentence-transformers==3.0.1
chromadb==0.5.5
fastapi==0.114.1
uvicorn==0.30.6
python-multipart==0.0.9
streamlit==1.38.0
python-dotenv==1.0.1
unstructured==0.15.0
unstructured[pdf]==0.15.0
pdfminer-six==20240706
numpy<2.0.0
requests==2.32.3
tiktoken==0.7.0
